LPS-78225 Create JournalHelper in API and implement it
