+++ modules/apps/journal-properties-transformer-listener/journal-properties-transformer-listener/src/main/java/com/liferay/journal/properties/transformer/listener/internal/JournalPropertiesTransformerListener.java
@@ -16,7 +16,7 @@ package com.liferay.journal.properties.transformer.listener.internal;
 
 import com.liferay.journal.constants.JournalPortletKeys;
 import com.liferay.journal.constants.JournalTransformerListenerKeys;
-import com.liferay.journal.util.impl.JournalUtil;
+import com.liferay.journal.util.JournalHelper;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -34,6 +34,7 @@ import java.util.Map;
 import java.util.Properties;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Brian Wing Shun Chan
@@ -101,7 +102,7 @@ public class JournalPropertiesTransformerListener
 			long articleGroupId = GetterUtil.getLong(
 				tokens.get("article_group_id"));
 
-			String script = JournalUtil.getTemplateScript(
+			String script = _journalHelper.getTemplateScript(
 				articleGroupId, _GLOBAL_PROPERTIES, newTokens, languageId);
 
 			PropertiesUtil.load(properties, script);
@@ -166,5 +167,8 @@ public class JournalPropertiesTransformerListener
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		JournalPropertiesTransformerListener.class);
+
+	@Reference
+	private JournalHelper _journalHelper;
 
 }
