+++ modules/apps/web-experience/journal/journal-service/src/main/java/com/liferay/journal/service/impl/JournalArticleLocalServiceImpl.java
@@ -71,12 +71,14 @@ import com.liferay.journal.model.JournalFolder;
 import com.liferay.journal.model.impl.JournalArticleDisplayImpl;
 import com.liferay.journal.service.base.JournalArticleLocalServiceBaseImpl;
 import com.liferay.journal.util.JournalConverter;
+import com.liferay.journal.util.JournalHelper;
 import com.liferay.journal.util.comparator.ArticleIDComparator;
 import com.liferay.journal.util.comparator.ArticleVersionComparator;
 import com.liferay.journal.util.impl.JournalUtil;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
 import com.liferay.petra.xml.XMLUtil;
+import com.liferay.portal.kernel.bean.BeanReference;
 import com.liferay.portal.kernel.comment.CommentManager;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
@@ -7861,7 +7863,7 @@ public class JournalArticleLocalServiceImpl
 
 			articleContent = articleDisplay.getContent();
 
-			articleDiffs = JournalUtil.diffHtml(
+			articleDiffs = _journalHelper.diffHtml(
 				article.getGroupId(), article.getArticleId(),
 				previousApprovedArticle.getVersion(), article.getVersion(),
 				LocaleUtil.toLanguageId(LocaleUtil.getSiteDefault()),
@@ -7995,7 +7997,7 @@ public class JournalArticleLocalServiceImpl
 			Hits hits = indexer.search(
 				searchContext, JournalUtil.SELECTED_FIELD_NAMES);
 
-			List<JournalArticle> articles = JournalUtil.getArticles(hits);
+			List<JournalArticle> articles = _journalHelper.getArticles(hits);
 
 			if (articles != null) {
 				return new BaseModelSearchResult<>(articles, hits.getLength());
@@ -8598,7 +8600,7 @@ public class JournalArticleLocalServiceImpl
 			long groupId, long folderId, String ddmStructureKey)
 		throws PortalException {
 
-		int restrictionType = JournalUtil.getRestrictionType(folderId);
+		int restrictionType = _journalHelper.getRestrictionType(folderId);
 
 		DDMStructure ddmStructure = ddmStructureLocalService.getStructure(
 			PortalUtil.getSiteGroupId(groupId),
@@ -8656,7 +8658,8 @@ public class JournalArticleLocalServiceImpl
 		}
 
 		if (Validator.isNotNull(layoutUuid)) {
-			Layout layout = JournalUtil.getArticleLayout(layoutUuid, groupId);
+			Layout layout = _journalHelper.getArticleLayout(
+				layoutUuid, groupId);
 
 			if (layout == null) {
 				throw new NoSuchLayoutException(
@@ -8955,6 +8958,9 @@ public class JournalArticleLocalServiceImpl
 
 	@ServiceReference(type = JournalFileUploadsConfiguration.class)
 	private JournalFileUploadsConfiguration _journalFileUploadsConfiguration;
+
+	@BeanReference(type = JournalHelper.class)
+	private JournalHelper _journalHelper;
 
 	private Date _previousCheckDate;
