+++ modules/apps/web-experience/journal/journal-service/src/main/java/com/liferay/journal/util/impl/JournalUtil.java
@@ -23,7 +23,6 @@ import com.liferay.journal.configuration.JournalServiceConfiguration;
 import com.liferay.journal.internal.transformer.JournalTransformer;
 import com.liferay.journal.internal.transformer.JournalTransformerListenerRegistryUtil;
 import com.liferay.journal.model.JournalArticle;
-import com.liferay.journal.model.JournalArticleDisplay;
 import com.liferay.journal.model.JournalFolder;
 import com.liferay.journal.model.JournalFolderConstants;
 import com.liferay.journal.model.JournalStructureConstants;
@@ -35,12 +34,9 @@ import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
 import com.liferay.petra.xml.XMLUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
-import com.liferay.portal.kernel.diff.CompareVersionsException;
-import com.liferay.portal.kernel.diff.DiffHtmlUtil;
 import com.liferay.portal.kernel.diff.DiffVersion;
 import com.liferay.portal.kernel.diff.DiffVersionsInfo;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.io.unsync.UnsyncStringReader;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -57,9 +53,6 @@ import com.liferay.portal.kernel.portlet.PortletProviderUtil;
 import com.liferay.portal.kernel.portlet.PortletRequestModel;
 import com.liferay.portal.kernel.portlet.ThemeDisplayModel;
 import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.Hits;
-import com.liferay.portal.kernel.search.Indexer;
-import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.service.ImageLocalServiceUtil;
 import com.liferay.portal.kernel.service.LayoutLocalServiceUtil;
 import com.liferay.portal.kernel.service.SubscriptionLocalServiceUtil;
@@ -72,7 +65,6 @@ import com.liferay.portal.kernel.templateparser.TransformerListener;
 import com.liferay.portal.kernel.theme.PortletDisplay;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.FriendlyURLNormalizerUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HtmlUtil;
@@ -94,7 +86,6 @@ import com.liferay.portal.kernel.xml.SAXReaderUtil;
 import com.liferay.portal.kernel.xml.XPath;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -244,141 +235,6 @@ public class JournalUtil {
 
 			stack.push(article);
 		}
-	}
-
-	public static String diffHtml(
-			long groupId, String articleId, double sourceVersion,
-			double targetVersion, String languageId,
-			PortletRequestModel portletRequestModel, ThemeDisplay themeDisplay)
-		throws Exception {
-
-		JournalArticle sourceArticle =
-			JournalArticleLocalServiceUtil.getArticle(
-				groupId, articleId, sourceVersion);
-
-		if (!JournalArticleLocalServiceUtil.isRenderable(
-				sourceArticle, portletRequestModel, themeDisplay)) {
-
-			throw new CompareVersionsException(sourceVersion);
-		}
-
-		JournalArticleDisplay sourceArticleDisplay =
-			JournalArticleLocalServiceUtil.getArticleDisplay(
-				sourceArticle, null, Constants.VIEW, languageId, 1,
-				portletRequestModel, themeDisplay);
-
-		JournalArticle targetArticle =
-			JournalArticleLocalServiceUtil.getArticle(
-				groupId, articleId, targetVersion);
-
-		if (!JournalArticleLocalServiceUtil.isRenderable(
-				targetArticle, portletRequestModel, themeDisplay)) {
-
-			throw new CompareVersionsException(targetVersion);
-		}
-
-		JournalArticleDisplay targetArticleDisplay =
-			JournalArticleLocalServiceUtil.getArticleDisplay(
-				targetArticle, null, Constants.VIEW, languageId, 1,
-				portletRequestModel, themeDisplay);
-
-		return DiffHtmlUtil.diff(
-			new UnsyncStringReader(sourceArticleDisplay.getContent()),
-			new UnsyncStringReader(targetArticleDisplay.getContent()));
-	}
-
-	public static String getAbsolutePath(
-			PortletRequest portletRequest, long folderId)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		if (folderId == JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
-			return themeDisplay.translate("home");
-		}
-
-		JournalFolder folder = JournalFolderLocalServiceUtil.getFolder(
-			folderId);
-
-		List<JournalFolder> folders = folder.getAncestors();
-
-		Collections.reverse(folders);
-
-		StringBundler sb = new StringBundler((folders.size() * 3) + 5);
-
-		sb.append(themeDisplay.translate("home"));
-		sb.append(StringPool.SPACE);
-
-		for (JournalFolder curFolder : folders) {
-			sb.append(StringPool.RAQUO_CHAR);
-			sb.append(StringPool.SPACE);
-			sb.append(curFolder.getName());
-		}
-
-		sb.append(StringPool.RAQUO_CHAR);
-		sb.append(StringPool.SPACE);
-		sb.append(folder.getName());
-
-		return sb.toString();
-	}
-
-	public static Layout getArticleLayout(String layoutUuid, long groupId) {
-		if (Validator.isNull(layoutUuid)) {
-			return null;
-		}
-
-		// The target page and the article must belong to the same group
-
-		Layout layout = LayoutLocalServiceUtil.fetchLayoutByUuidAndGroupId(
-			layoutUuid, groupId, false);
-
-		if (layout == null) {
-			layout = LayoutLocalServiceUtil.fetchLayoutByUuidAndGroupId(
-				layoutUuid, groupId, true);
-		}
-
-		return layout;
-	}
-
-	/**
-	 * @deprecated As of 4.0.0, with no direct replacement
-	 */
-	@Deprecated
-	public static List<JournalArticle> getArticles(Hits hits)
-		throws PortalException {
-
-		List<com.liferay.portal.kernel.search.Document> documents =
-			hits.toList();
-
-		List<JournalArticle> articles = new ArrayList<>(documents.size());
-
-		for (com.liferay.portal.kernel.search.Document document : documents) {
-			String articleId = document.get(Field.ARTICLE_ID);
-			long groupId = GetterUtil.getLong(
-				document.get(Field.SCOPE_GROUP_ID));
-
-			JournalArticle article =
-				JournalArticleLocalServiceUtil.fetchLatestArticle(
-					groupId, articleId, WorkflowConstants.STATUS_APPROVED);
-
-			if (article == null) {
-				articles = null;
-
-				Indexer<JournalArticle> indexer =
-					IndexerRegistryUtil.getIndexer(JournalArticle.class);
-
-				long companyId = GetterUtil.getLong(
-					document.get(Field.COMPANY_ID));
-
-				indexer.delete(companyId, document.getUID());
-			}
-			else if (articles != null) {
-				articles.add(article);
-			}
-		}
-
-		return articles;
 	}
 
 	public static DiffVersionsInfo getDiffVersionsInfo(
@@ -554,28 +410,6 @@ public class JournalUtil {
 		}
 
 		return themeDisplay.getPlid();
-	}
-
-	public static int getRestrictionType(long folderId) {
-		int restrictionType = JournalFolderConstants.RESTRICTION_TYPE_INHERIT;
-
-		JournalFolder folder = JournalFolderLocalServiceUtil.fetchFolder(
-			folderId);
-
-		if (folder != null) {
-			restrictionType = folder.getRestrictionType();
-		}
-
-		return restrictionType;
-	}
-
-	public static String getTemplateScript(
-			long groupId, String ddmTemplateKey, Map<String, String> tokens,
-			String languageId)
-		throws PortalException {
-
-		return _getTemplateScript(
-			groupId, ddmTemplateKey, tokens, languageId, true);
 	}
 
 	public static Map<String, String> getTokens(
