+++ modules/apps/web-experience/journal/journal-service/src/main/resources/META-INF/spring/ext-spring.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+
+<beans
+	default-destroy-method="destroy"
+	default-init-method="afterPropertiesSet"
+	xmlns="http://www.springframework.org/schema/beans"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
+>
+	<bean class="com.liferay.journal.internal.util.JournalHelperUtil" id="com.liferay.journal.internal.util.JournalHelperUtil">
+		<property name="journalHelper" ref="com.liferay.journal.util.JournalHelper" />
+	</bean>
+	<bean class="com.liferay.journal.util.impl.JournalHelperImpl" id="com.liferay.journal.util.JournalHelper" />
+</beans>
